+++ SoA/ChunkGridRenderStage.cpp
@@ -86,7 +86,7 @@ void ChunkGridRenderStage::render() {
 
     f32v3 posOffset;
     // TODO(Ben): Got a race condition here that causes rare crash
-    for (i32 i = 0; i < chunks.size(); i++) {
+    for (size_t i = 0; i < chunks.size(); i++) {
         const Chunk* chunk = chunks[i];
         posOffset = f32v3(f64v3(chunk->voxelPosition) - m_gameRenderParams->chunkCamera->getPosition());
