+++ SoA/InputMapper.cpp
@@ -31,7 +31,7 @@ InputMapper::~InputMapper() {
 
 bool InputMapper::getInputState(const InputID id) {
     // Check Input
-    if (id < 0 || id >= m_inputs.size()) return false;
+    if (id < 0 || id >= (int)m_inputs.size()) return false;
     return m_keyStates[m_inputs.at(id).key];
 }
 
@@ -143,7 +143,7 @@ void InputMapper::saveInputs(const nString &filePath /* = INPUTMAPPER_DEFAULT_CO
 }
 
 VirtualKey InputMapper::getKey(const InputID id) {
-    if (id < 0 || id >= m_inputs.size()) return VKEY_HIGHEST_VALUE;
+    if (id < 0 || id >= (int)m_inputs.size()) return VKEY_HIGHEST_VALUE;
     return m_inputs.at(id).key;
 }
 
@@ -152,7 +152,7 @@ void InputMapper::setKey(const InputID id, VirtualKey key) {
     VirtualKey oldKey = m_inputs.at(id).key;
     auto& it = m_keyCodeMap.find(oldKey);
     auto& vec = it->second;
-    for (int i = 0; i < vec.size(); i++) {
+    for (size_t i = 0; i < vec.size(); i++) {
         // Remove the input from the vector keyed on VirtualKey
         if (vec[i] == id) {
             vec[i] = vec.back();
@@ -166,7 +166,7 @@ void InputMapper::setKey(const InputID id, VirtualKey key) {
 }
 
 void InputMapper::setKeyToDefault(const InputID id) {
-    if (id < 0 || id >= m_inputs.size()) return;
+    if (id < 0 || id >= (int)m_inputs.size()) return;
     setKey(id, m_inputs.at(id).defaultKey);
 }
