+++ SoA/MainMenuSystemViewer.cpp
@@ -50,7 +50,7 @@ void MainMenuSystemViewer::update() {
     const f32 HOVER_SPEED = 0.08f;
     const f32 HOVER_SIZE_INC = 7.0f;
 
-    m_camera->setClippingPlane(0.1f * KM_PER_M, m_camera->getFarClip());
+    m_camera->setClippingPlane((f32)(0.1 * KM_PER_M), m_camera->getFarClip());
     // Target closest point on sphere
     m_camera->setTargetFocalPoint(getTargetPosition() -
                                   f64v3(glm::normalize(m_camera->getDirection())) * getTargetRadius());
@@ -62,8 +62,8 @@ void MainMenuSystemViewer::update() {
 
         f64v3 relativePos = it.second.position - m_camera->getPosition();
         f64 distance = glm::length(relativePos);
-        f32 radiusPixels;
-        f32 radius;
+        f64 radiusPixels;
+        f64 radius;
 
         BodyArData& data = bodyArData[it.first];
         f32 hoverTime = data.hoverTime;
@@ -83,16 +83,16 @@ void MainMenuSystemViewer::update() {
                 // Get radius of projected sphere
                 radius = m_spaceSystem->m_sphericalGravityCT.get(componentID).radius;
                 radiusPixels = (radius /
-                                (tan(m_camera->getFieldOfView() / 2) * distance)) *
-                                (m_viewport.y / 2.0f);
+                                (tan((f64)m_camera->getFieldOfView() / 2.0) * distance)) *
+                                ((f64)m_viewport.y / 2.0);
             } else {
-                radius = 1000.0f;
+                radius = 1000.0;
                 radiusPixels = (radius /
-                                (tan(m_camera->getFieldOfView() / 2) * distance)) *
-                                (m_viewport.y / 2.0f);
+                                (tan((f64)m_camera->getFieldOfView() / 2.0) * distance)) *
+                                ((f64)m_viewport.y / 2.0);
             }
 
-            f32 selectorSize = radiusPixels * 2.0f + 3.0f;
+            f32 selectorSize = (f32)(radiusPixels * 2.0 + 3.0);
             if (selectorSize < MIN_SELECTOR_SIZE) selectorSize = MIN_SELECTOR_SIZE;
 
             // Interpolate size
@@ -150,8 +150,8 @@ nString MainMenuSystemViewer::getTargetName() {
 }
 
 void MainMenuSystemViewer::onMouseButtonDown(Sender sender, const vui::MouseButtonEvent& e) {
-    m_mouseCoords.x = e.x;
-    m_mouseCoords.y = e.y;
+    m_mouseCoords.x = (f32)e.x;
+    m_mouseCoords.y = (f32)e.y;
     if (e.button == vui::MouseButton::LEFT) {
         mouseButtons[0] = true;
         // Target a body if we clicked on one
@@ -183,8 +183,8 @@ void MainMenuSystemViewer::onMouseButtonDown(Sender sender, const vui::MouseButt
 }
 
 void MainMenuSystemViewer::onMouseButtonUp(Sender sender, const vui::MouseButtonEvent& e) {
-    m_mouseCoords.x = e.x;
-    m_mouseCoords.y = e.y;
+    m_mouseCoords.x = (f32)e.x;
+    m_mouseCoords.y = (f32)e.y;
     if (e.button == vui::MouseButton::LEFT) {
         mouseButtons[0] = false;
     } else {
@@ -194,15 +194,15 @@ void MainMenuSystemViewer::onMouseButtonUp(Sender sender, const vui::MouseButton
 
 void MainMenuSystemViewer::onMouseWheel(Sender sender, const vui::MouseWheelEvent& e) {
 #define SCROLL_SPEED 0.1f
-    m_camera->offsetTargetFocalLength(m_camera->getTargetFocalLength() * SCROLL_SPEED * -e.dy);
+    m_camera->offsetTargetFocalLength((f32)m_camera->getTargetFocalLength() * SCROLL_SPEED * -e.dy);
     if (m_camera->getTargetFocalLength() < 0.1f) {
         m_camera->setTargetFocalLength(0.1f);
     }
 }
 
 void MainMenuSystemViewer::onMouseMotion(Sender sender, const vui::MouseMotionEvent& e) {
-    m_mouseCoords.x = e.x;
-    m_mouseCoords.y = e.y;
+    m_mouseCoords.x = (f32)e.x;
+    m_mouseCoords.y = (f32)e.y;
 
 #define MOUSE_SPEED 0.1f
     if (mouseButtons[0]) {
@@ -230,7 +230,7 @@ void MainMenuSystemViewer::pickStartLocation(vecs::EntityID eid) {
 
     cid = m_spaceSystem->m_sphericalGravityCT.getComponentID(eid);
     if (!cid) return;
-    f32 radius = m_spaceSystem->m_sphericalGravityCT.get(cid).radius;
+    f32 radius = (f32)m_spaceSystem->m_sphericalGravityCT.get(cid).radius;
 
     // Compute the intersection
     f32v3 normal, hitpoint;
@@ -300,6 +300,6 @@ void MainMenuSystemViewer::computeGridPosition(const f32v3& hitpoint, f32 radius
     VoxelPosition2D facePos;
     facePos.pos = f64v2(gridHit);
     facePos.face = m_selectedCubeFace;
-    height = m_spaceSystem->m_sphericalTerrainCT.getFromEntity(m_targetEntity).cpuGenerator->getTerrainHeight(facePos) * KM_PER_VOXEL;
+    height = (f32)(m_spaceSystem->m_sphericalTerrainCT.getFromEntity(m_targetEntity).cpuGenerator->getTerrainHeight(facePos) * KM_PER_VOXEL);
     m_selectedGridPos.y = height;
 }
